changed schema and made more modular

+++ src/main/java/io/d8a/conjure/DoubleMetricNode.java
@@ -17,7 +17,7 @@ public class DoubleMetricNode extends MetricNode implements ConjureTemplateNode
   @Override
   public Double getValue()
   {
-    return counter.nextValue()+0.1234d;
+    return (double)counter.nextValue()+0.1234d;
   }
 
   @Override
